+++ service/controller/app/resources.go
@@ -38,6 +38,7 @@ type appResourcesConfig struct {
 	HTTPClientTimeout time.Duration
 	ImageRegistry     string
 	UniqueApp         bool
+	WebhookBaseURL    string
 }
 
 func newAppResources(config appResourcesConfig) ([]resource.Interface, error) {
@@ -63,6 +64,9 @@ func newAppResources(config appResourcesConfig) ([]resource.Interface, error) {
 	}
 	if config.ImageRegistry == "" {
 		return nil, microerror.Maskf(invalidConfigError, "%T.ImageRegistry must not be empty", config)
+	}
+	if config.WebhookBaseURL == "" {
+		return nil, microerror.Maskf(invalidConfigError, "%T.WebhookBaseURL must not be empty", config)
 	}
 
 	var valuesService *values.Values
@@ -123,6 +127,7 @@ func newAppResources(config appResourcesConfig) ([]resource.Interface, error) {
 			Logger: config.Logger,
 
 			ChartNamespace: config.ChartNamespace,
+			WebhookBaseURL: config.WebhookBaseURL,
 		}
 
 		ops, err := chart.New(c)